+++ CPU_Renderer/CPU_Renderer.vcxproj
@@ -175,7 +175,13 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Source\Colors\Colors.cpp" />
+    <ClCompile Include="Source\Display\Display.cpp" />
     <ClCompile Include="Source\Main.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Source\Colors\Colors.h" />
+    <ClInclude Include="Source\Display\Display.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
