+++ CPU_Renderer/Source/Colors/Colors.h
@@ -1,0 +1,15 @@
+
+#ifndef COLORS_H
+#define COLORS_H
+
+#include <stdint.h>
+
+extern uint32_t ALICE_BLUE;
+extern uint32_t BLACK;
+extern uint32_t RED;
+extern uint32_t GREEN;
+extern uint32_t BLUE;
+extern uint32_t YELLOW;
+extern uint32_t WHITE;
+
+#endif
